server_backup/pyproject.toml
[project]
name = "ops-agent"
version = "0.1.0"
description = "Ops AI Agent - Event-driven operations AI system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.7.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "redis>=5.0.0",
    "structlog>=24.1.0",
    "orjson>=3.10.0",
    "pyjwt>=2.10.0",
    "cryptography>=44.0.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "asyncssh>=2.18.0",
    "httpx>=0.28.0",
    "sse-starlette>=2.2.0",
    "pgvector>=0.3.0",
    "poethepoet>=0.42.1",
    "python-multipart>=0.0.22",
    "docling>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "websockets>=14.0",
    "aiomysql>=0.2.0",
    "motor>=3.6.0",
    "rich>=13.0.0",
    "pyhive>=0.7.0",
    "thrift>=0.22.0",
    "docker>=7.0.0",
    "dashscope>=1.25.15",
    "av>=17.0.0",
    "regex>=2024.0.0",
    "apscheduler>=3.10.0",
    "bcrypt>=4.2.0",
]

[project.optional-dependencies]

[tool.ruff]
line-length = 100

[tool.poe.tasks]
dev = "uvicorn src.main:app --reload --host 0.0.0.0 --port 8000"
lint = "ruff check src/"
format = "ruff format src/"
migrate = "alembic upgrade head"
"migrate:new" = "alembic revision --autogenerate"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "agent: agent integration tests (require LLM + docker services)",
]

[tool.poe.tasks.test]
cmd = "pytest tests/ -x -v --ignore=tests/agent"
[tool.poe.tasks."test:api"]
cmd = "pytest tests/api/ -x -v"
[tool.poe.tasks."test:unit"]
cmd = "pytest tests/unit/ -x -v"
[tool.poe.tasks."test:agent"]
cmd = "pytest tests/agent/ -x -v -m agent"

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "pytest>=9.0.2",
    "pytest-playwright>=0.7.2",
    "pytest-asyncio>=1.3.0",
    "pytest-timeout>=2.3.0",
    "asgi-lifespan>=2.0.0",
    "pytest-cov>=6.0.0",
]
